Refactor marketing overview components
