+++ apps/cms/src/app/cms/marketing/components/MarketingOverviewHero.tsx
@@ -1,0 +1,13 @@
+export function MarketingOverviewHero() {
+  return (
+    <header className="space-y-2">
+      <h1 className="text-2xl font-semibold text-foreground">Marketing workspace</h1>
+      <p className="max-w-2xl text-sm text-muted-foreground">
+        Launch campaigns, manage incentives, and keep sales aligned with the latest customer signals.
+        Start with a guided tool and revisit recent activity across shops.
+      </p>
+    </header>
+  );
+}
+
+export default MarketingOverviewHero;
