+++ apps/cms/src/app/cms/marketing/components/MarketingSummaryCards.tsx
@@ -1,0 +1,17 @@
+import { AnalyticsSummaryCard, type AnalyticsSummaryCardProps } from "@ui/components/cms/marketing";
+
+export interface MarketingSummaryCardsProps {
+  cards: AnalyticsSummaryCardProps[];
+}
+
+export function MarketingSummaryCards({ cards }: MarketingSummaryCardsProps) {
+  return (
+    <section className="grid gap-4 lg:grid-cols-2">
+      {cards.map((card) => (
+        <AnalyticsSummaryCard key={card.title} {...card} />
+      ))}
+    </section>
+  );
+}
+
+export default MarketingSummaryCards;
